Reorganise models in the `core` app
- mixins, pages, snippets in separate modules in `core/models/`
- move the various classes in `utils/models.py` to the relevant module
  in `core/models/` and update references to these classes throughout codebase
- migrations in separate commit
