+++ wagtailio/core/models/__init__.py
@@ -1,0 +1,3 @@
+from .mixins import *  # noqa
+from .pages import *  # noqa
+from .snippets import *  # noqa
