+++ wagtailio/core/models/snippets.py
@@ -2,14 +2,8 @@ from django.db import models
 
 from modelcluster.fields import ParentalKey
 from modelcluster.models import ClusterableModel
-from wagtail.admin.edit_handlers import (
-    FieldPanel,
-    InlinePanel,
-    MultiFieldPanel,
-    PageChooserPanel,
-)
+from wagtail.admin.edit_handlers import FieldPanel, InlinePanel, PageChooserPanel
 from wagtail.core.models import Orderable
-from wagtail.images.edit_handlers import ImageChooserPanel
 from wagtail.snippets.models import register_snippet
 
 
@@ -84,56 +78,3 @@ class MenuSnippet(ClusterableModel):
 
 
 register_snippet(MenuSnippet)
-
-
-class SocialMediaMixin(models.Model):
-    social_text = models.CharField(
-        "Meta description",
-        max_length=255,
-        blank=True,
-        help_text="Description of this page as it should appear when shared on social networks, or in Google results",
-    )
-    social_image = models.ForeignKey(
-        "images.WagtailIOImage",
-        models.SET_NULL,
-        verbose_name="Meta image",
-        null=True,
-        blank=True,
-        related_name="+",
-        help_text="Image to appear alongside 'Meta description', particularly for sharing on social networks",
-    )
-
-    panels = [
-        MultiFieldPanel(
-            [FieldPanel("social_text"), ImageChooserPanel("social_image")],
-            "Social/Meta descriptions",
-        )
-    ]
-
-    class Meta:
-        abstract = True
-
-
-class CrossPageMixin(models.Model):
-    listing_image = models.ForeignKey(
-        "images.WagtailIOImage",
-        models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name="+",
-        help_text="Image to display along with summary, when this page is linked from elsewhere in the site.",
-    )
-    listing_intro = models.TextField(
-        blank=True,
-        help_text="Summary of this page to display when this is linked from elsewhere in the site.",
-    )
-
-    panels = [
-        MultiFieldPanel(
-            [FieldPanel("listing_intro"), ImageChooserPanel("listing_image")],
-            "Cross-page behaviour",
-        )
-    ]
-
-    class Meta:
-        abstract = True
